feat: move db servie to services
