+++ app/routes/manifest.py
@@ -4,7 +4,7 @@ from quart import Blueprint
 
 import config
 
-from ..db.db import get_user
+from ..services.db import get_user
 from .utils import respond_with
 
 manifest_blueprint = Blueprint("manifest", __name__)
